+++ infra/core/psn/spoke/prod/locals.tf
@@ -6,8 +6,6 @@ locals {
     instance_number = "01"
   }
 
-  project = "${local.environment.prefix}-${local.environment.environment}-${local.environment.location}-%s-${local.environment.instance_number}"
-
   networking = {
     hub = {
       resource_group_name = "pagopa-rg-hub-network-italynorth"