+++ infra/resources/_modules/function_apps/locals.tf
@@ -79,7 +79,6 @@ locals {
     "generateEntityConfiguration",
     "sendEmailOnWalletInstanceCreation",
     "sendEmailOnWalletInstanceRevocation",
-    "migrateWalletInstances"
   ]
 
   function_app_user = {
@@ -131,8 +130,6 @@ locals {
         PidIssuerApiClientPrivateKey  = "@Microsoft.KeyVault(VaultName=${var.key_vault_wallet_name};SecretName=PidIssuerApiClientPrivateKeyPre)"
         GoogleAppCredentialsEncoded   = "@Microsoft.KeyVault(VaultName=${var.key_vault_wallet_name};SecretName=GoogleAppCredentialsEncoded)"
         AllowedDeveloperUsers         = "@Microsoft.KeyVault(VaultName=${var.key_vault_wallet_name};SecretName=AllowedDeveloperUsers)"
-        StorageConnectionString       = var.is_psn ? null : "@Microsoft.KeyVault(VaultName=${var.key_vault_wallet_name};SecretName=StorageUatConnectionString)"
-        CosmosAccountConnectionString = "@Microsoft.KeyVault(VaultName=${var.key_vault_wallet_name};SecretName=CosmosAccountConnectionString)"
       }
     )
   }